+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+### Added
+- Add new `WebhookAttemptsEndpoint` and `WebhookAttempt` resource
+
+### Modify
+- Add `Webhook.status` 
+
 ## [1.0.1] - 2021-06-03
 ### Fixed
 - Fix nullable types on EnvelopeTag